requirements.txt
discord.py
datetime
asyncio